Free unit tests for Mac: BookmarkHTMLWriterTest, BookmarkUtilsTest, DownloadRequestManagerTest, URLFetcher*Test, TimeFormat, RenderWidgetTest
Review URL: http://codereview.chromium.org/50081

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@12269 0039d316-1c4b-4281-b951-d872f2087c98

+++ chrome/browser/bookmarks/bookmark_utils_unittest.cc
@@ -22,19 +22,19 @@ TEST_F(BookmarkUtilsTest, GetBookmarksContainingText) {
 
   std::vector<BookmarkNode*> nodes;
   bookmark_utils::GetBookmarksContainingText(&model, L"foo", 100, &nodes);
-  ASSERT_EQ(1, nodes.size());
+  ASSERT_EQ(static_cast<size_t>(1), nodes.size());
   EXPECT_TRUE(nodes[0] == n1);
   EXPECT_TRUE(bookmark_utils::DoesBookmarkContainText(n1, L"foo"));
   nodes.clear();
 
   bookmark_utils::GetBookmarksContainingText(&model, L"cnn", 100, &nodes);
-  ASSERT_EQ(1, nodes.size());
+  ASSERT_EQ(static_cast<size_t>(1), nodes.size());
   EXPECT_TRUE(nodes[0] == n2);
   EXPECT_TRUE(bookmark_utils::DoesBookmarkContainText(n2, L"cnn"));
   nodes.clear();
 
   bookmark_utils::GetBookmarksContainingText(&model, L"foo bar", 100, &nodes);
-  ASSERT_EQ(1, nodes.size());
+  ASSERT_EQ(static_cast<size_t>(1), nodes.size());
   EXPECT_TRUE(nodes[0] == n1);
   EXPECT_TRUE(bookmark_utils::DoesBookmarkContainText(n1, L"foo bar"));
   nodes.clear();
